+++ pkg/cover/backend/gvisor.go
@@ -49,7 +49,7 @@ func makeGvisor(target *targets.Target, objDir, srcDir, buildDir string, modules
 	impl := &Impl{
 		Units:  units,
 		Frames: frames,
-		RestorePC: func(pc uint32) uint64 {
+		RestorePC: func(pc uint64) uint64 {
 			return uint64(pc)
 		},
 	}
